delete x/stream
