+++ cmd/client/flags.go
@@ -12,7 +12,6 @@ import (
 	"github.com/okex/exchain/x/common/analyzer"
 	evmtypes "github.com/okex/exchain/x/evm/types"
 	"github.com/okex/exchain/x/evm/watcher"
-	"github.com/okex/exchain/x/stream"
 	"github.com/okex/exchain/x/token"
 	"github.com/spf13/cobra"
 	tmdb "github.com/tendermint/tm-db"
@@ -24,13 +23,13 @@ func RegisterAppFlag(cmd *cobra.Command) {
 	cmd.Flags().Bool(rpc.FlagPersonalAPI, true, "Enable the personal_ prefixed set of APIs in the Web3 JSON-RPC spec")
 	cmd.Flags().Bool(evmtypes.FlagEnableBloomFilter, false, "Enable bloom filter for event logs")
 	cmd.Flags().Int64(filters.FlagGetLogsHeightSpan, 2000, "config the block height span for get logs")
-	cmd.Flags().String(stream.NacosTmrpcUrls, "", "Stream plugin`s nacos server urls for discovery service of tendermint rpc")
-	cmd.Flags().MarkHidden(stream.NacosTmrpcUrls)
-	cmd.Flags().String(stream.NacosTmrpcNamespaceID, "", "Stream plugin`s nacos namepace id for discovery service of tendermint rpc")
-	cmd.Flags().MarkHidden(stream.NacosTmrpcNamespaceID)
-	cmd.Flags().String(stream.NacosTmrpcAppName, "", "Stream plugin`s tendermint rpc name in eureka or nacos")
-	cmd.Flags().MarkHidden(stream.NacosTmrpcAppName)
-	cmd.Flags().String(stream.RpcExternalAddr, "127.0.0.1:26657", "Set the rpc-server external ip and port, when it is launched by Docker (default \"127.0.0.1:26657\")")
+	cmd.Flags().String(rpc.FlagNacosTmrpcUrls, "", "Stream plugin`s nacos server urls for discovery service of tendermint rpc")
+	cmd.Flags().MarkHidden(rpc.FlagNacosTmrpcUrls)
+	cmd.Flags().String(rpc.FlagNacosTmrpcNamespaceID, "", "Stream plugin`s nacos namepace id for discovery service of tendermint rpc")
+	cmd.Flags().MarkHidden(rpc.FlagNacosTmrpcNamespaceID)
+	cmd.Flags().String(rpc.FlagNacosTmrpcAppName, "", "Stream plugin`s tendermint rpc name in eureka or nacos")
+	cmd.Flags().MarkHidden(rpc.FlagNacosTmrpcAppName)
+	cmd.Flags().String(rpc.FlagRpcExternalAddr, "127.0.0.1:26657", "Set the rpc-server external ip and port, when it is launched by Docker (default \"127.0.0.1:26657\")")
 
 	cmd.Flags().String(rpc.FlagRateLimitAPI, "", "Set the RPC API to be controlled by the rate limit policy, such as \"eth_getLogs,eth_newFilter,eth_newBlockFilter,eth_newPendingTransactionFilter,eth_getFilterChanges\"")
 	cmd.Flags().Int(rpc.FlagRateLimitCount, 0, "Set the count of requests allowed per second of rpc rate limiter")
