--- x/stream/alias.go
@@ -1,13 +1,0 @@
-package stream
-
-import "github.com/okex/exchain/x/stream/types"
-
-const (
-	ModuleName   = types.ModuleName
-	QuerierRoute = types.QuerierRoute
-	RouterKey    = types.RouterKey
-)
-
-type (
-	DexKeeper = types.DexKeeper
-)
