--- x/stream/module.go
@@ -1,105 +1,0 @@
-package stream
-
-import (
-	"encoding/json"
-
-	"github.com/okex/exchain/libs/cosmos-sdk/client/context"
-	"github.com/okex/exchain/libs/cosmos-sdk/codec"
-	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
-	"github.com/okex/exchain/libs/cosmos-sdk/types/module"
-	"github.com/gorilla/mux"
-	"github.com/spf13/cobra"
-	abci "github.com/okex/exchain/libs/tendermint/abci/types"
-)
-
-// type check to ensure the interface is properly implemented
-var (
-	_ module.AppModule      = AppModule{}
-	_ module.AppModuleBasic = AppModuleBasic{}
-)
-
-// app module Basics object
-type AppModuleBasic struct{}
-
-func (AppModuleBasic) Name() string {
-	return ModuleName
-}
-
-func (AppModuleBasic) RegisterCodec(cdc *codec.Codec) {
-}
-
-func (AppModuleBasic) DefaultGenesis() json.RawMessage {
-	return nil
-}
-
-// Validation check of the Genesis
-func (AppModuleBasic) ValidateGenesis(bz json.RawMessage) error {
-	return nil
-}
-
-// Register rest routes
-func (AppModuleBasic) RegisterRESTRoutes(ctx context.CLIContext, rtr *mux.Router) {
-}
-
-// Get the root query command of this module
-func (AppModuleBasic) GetQueryCmd(cdc *codec.Codec) *cobra.Command {
-	return nil
-}
-
-// Get the root tx command of this module
-func (AppModuleBasic) GetTxCmd(cdc *codec.Codec) *cobra.Command {
-	return nil
-}
-
-type AppModule struct {
-	AppModuleBasic
-	keeper Keeper
-}
-
-// NewAppModule creates a new AppModule Object
-func NewAppModule(k Keeper) AppModule {
-	return AppModule{
-		AppModuleBasic: AppModuleBasic{},
-		keeper:         k,
-	}
-}
-
-func (AppModule) Name() string {
-	return ModuleName
-}
-
-func (am AppModule) RegisterInvariants(ir sdk.InvariantRegistry) {}
-
-func (am AppModule) Route() string {
-	return RouterKey
-}
-
-func (am AppModule) NewHandler() sdk.Handler {
-	return nil
-}
-func (am AppModule) QuerierRoute() string {
-	return QuerierRoute
-}
-
-func (am AppModule) NewQuerierHandler() sdk.Querier {
-	return func(ctx sdk.Context, path []string, req abci.RequestQuery) (res []byte, err sdk.Error) {
-		return nil, nil
-	}
-}
-
-func (am AppModule) BeginBlock(ctx sdk.Context, _ abci.RequestBeginBlock) {
-	BeginBlocker(ctx, am.keeper)
-}
-
-func (am AppModule) EndBlock(ctx sdk.Context, _ abci.RequestEndBlock) []abci.ValidatorUpdate {
-	EndBlocker(ctx, am.keeper)
-	return nil
-}
-
-func (am AppModule) InitGenesis(ctx sdk.Context, data json.RawMessage) []abci.ValidatorUpdate {
-	return nil
-}
-
-func (am AppModule) ExportGenesis(ctx sdk.Context) json.RawMessage {
-	return nil
-}
