--- x/stream/types/keys.go
@@ -1,8 +1,0 @@
-package types
-
-const (
-	ModuleName   = "stream"
-	QuerierRoute = ModuleName
-	// RouterKey is the msg router key for the backend module
-	RouterKey = ""
-)
